CTOO8: Restore changes done to cardtable, moved List table to own folder
