+++ components/src/core/components/CardTable/Decorator/decorator-mixin.ts
@@ -1,7 +1,7 @@
 import {defineComponent} from 'vue';
 
 export const decoratorMixin = defineComponent({
-  inject: ['tableProps'],
+  inject: ['tableProps', 'screenState'],
 });
 
 export default decoratorMixin;
