+++ components/src/core/components/ListTable/list-table.scss
@@ -1,0 +1,49 @@
+@import '../CardTable/Decorator/card.scss';
+@import '../CardTable/Header/default-header.scss';
+
+.empty-msg-container {
+  vertical-align: top;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  height: calc(100vh - 230px);
+  .empty-msg {
+    margin-top: 0.5rem;
+    margin-bottom: 0.5rem;
+    display: flex;
+    align-items: center;
+    flex-direction: column;
+    .caption {
+      margin-top: 0.5rem;
+      margin-bottom: 0.5rem;
+      text-align: center;
+      color: #929baa;
+    }
+    .icon {
+      width: 100% !important;
+      height: unset !important;
+    }
+  }
+}
+
+::v-deep(.oxd-table-body) {
+  grid-gap: unset !important;
+}
+::v-deep(.oxd-table-header) {
+  padding: 10px 0;
+  border-bottom: 2px solid #dee2e6;
+}
+::v-deep(.oxd-table-card) {
+  border-radius: 0 !important;
+  &:not(:first-child) {
+    border-top: 1px solid #e8eaef;
+  }
+}
+::v-deep(.oxd-table-cell-actions) {
+  flex-wrap: nowrap;
+}
+::v-deep(.oxd-table-cell-actions .oxd-select-text) {
+  max-width: 195px;
+  width: fit-content;
+}
